Desktop/Heart_Guard-main/IntroECG_master/EchoNext_Minimodel/flask_app/requirements.txt
flask
torch
numpy
pandas
scipy
Pillow
xmltodict
pytorch-lightning
